Update AltoRouter.php
Just to be perfectionnist,

This line should be in the example:
```php
		// Force request_order to be GP
		// http://www.mail-archive.com/internals@lists.php.net/msg33119.html
		$_REQUEST = array_merge($_GET, $_POST);
```

Because AltoRouter manages routes. It's not a Request manager object. So it's better if developper who want to integrate AltoRouter learns this PHP issue by reading the AltoRouter example... He can consequently keep it in mind when he's writing his request strategy.